video-analysis/requirements.txt
moviepy==1.0.3
openai-whisper==20231117
transformers==4.38.2
torch==2.2.1
soundfile==0.12.1
numpy==1.26.4
matplotlib==3.8.2
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7
pydantic==2.6.1
assemblyai==0.17.0
python-dotenv==1.0.0
streamlit==1.44.1
watchdog==4.0.0
plotly==5.18.0
pandas==2.1.1
opencv-python==4.8.1.78
nltk==3.8.1
anthropic==0.50.0
nrclex==3.0.0
statsmodels==0.14.1
textblob==0.17.1
